requirements.txt
streamlit
pandas
sentence-transformers
openai
python-dotenv
google-generativeai
pydeck
faiss